requirements.txt
asgiref==3.5.2
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-advanced-password-validation==1.1.1
django-allauth==0.51.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rest-condition==1.0.3
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
wincertstore==0.2
